requirements.txt
kalshi_python_sync
python-dotenv
requests
